==> 01 Data exploration.ipynb <==
# Data exploration

import pandas as pd
import numpy as np
import seaborn as sns

dataset = pd.read_csv('train.csv')
testset = pd.read_csv('test.csv')

dataset.head()

dataset.info()

testset.info()

As it can be seen from the info-listing above, features Alley, FireplaceQu, PoolQC, Fence, MiscFeature contain a huge amount of non-null values. Let's remove them entirely. Also for MasVnrArea feature, over half of the values are zeroes, so let's remove it as well.

dataset = dataset.drop(['Alley', 'FireplaceQu', 'PoolQC', 'Fence', 'MiscFeature', 'MasVnrArea'], axis=1)
testset = testset.drop(['Alley', 'FireplaceQu', 'PoolQC', 'Fence', 'MiscFeature', 'MasVnrArea'], axis=1)

From features that contain null values, BsmtFinSF1, BsmtFinSF2, BsmtFullBath, BsmtHalfBath, BsmtUnfSF, GarageArea, GarageCars, GarageYrBlt, LotFrontage, MasVnrArea, TotalBsmtSF are numerical values, let's handle them first.

dataset['LotFrontage'].describe()

testset['LotFrontage'].describe()

dataset['GarageYrBlt'].describe()

testset['GarageYrBlt'].describe()

testset['BsmtFinSF2'].describe()

testset['BsmtFullBath'].describe()

testset['BsmtHalfBath'].describe()

testset['GarageCars'].describe()

For GarageYrBlt (the year garage was built), BsmtFinSF2, BsmtFullBath, BsmtHalfBath, GarageCars features, mode is the best value to replace the null values, instead of mean or median.

dataset['GarageYrBlt'] = dataset['GarageYrBlt'].fillna(dataset['GarageYrBlt'].mode()[0])
testset['GarageYrBlt'] = testset['GarageYrBlt'].fillna(testset['GarageYrBlt'].mode()[0])
testset['BsmtFinSF2'] = testset['BsmtFinSF2'].fillna(testset['BsmtFinSF2'].mode()[0])
testset['BsmtFullBath'] = testset['BsmtFullBath'].fillna(testset['BsmtFullBath'].mode()[0])
testset['BsmtHalfBath'] = testset['BsmtHalfBath'].fillna(testset['BsmtHalfBath'].mode()[0])
testset['GarageCars'] = testset['GarageCars'].fillna(testset['GarageCars'].mode()[0])

testset['BsmtFinSF1'].describe()

testset['BsmtUnfSF'].describe()

testset['TotalBsmtSF'].describe()

testset['GarageArea'].describe()

For LotFrontage, BsmtFinSF1, BsmtUnfSF, TotalBsmtSF and GarageArea mean is a good choice to replace the missing values.

dataset['LotFrontage'] = dataset['LotFrontage'].fillna(dataset['LotFrontage'].mean())
testset['LotFrontage'] = testset['LotFrontage'].fillna(testset['LotFrontage'].mean())
testset['BsmtFinSF1'] = testset['BsmtFinSF1'].fillna(testset['BsmtFinSF1'].mean())
testset['BsmtUnfSF'] = testset['BsmtUnfSF'].fillna(testset['BsmtUnfSF'].mean())
testset['TotalBsmtSF'] = testset['TotalBsmtSF'].fillna(testset['TotalBsmtSF'].mean())
testset['GarageArea'] = testset['GarageArea'].fillna(testset['GarageArea'].mean())

dataset.info()

testset.info()

dataset['BsmtQual']

Let's replace all the empty categorigal values with the mode of that feature.

dataset['MasVnrType'] = dataset['MasVnrType'].fillna(dataset['MasVnrType'].mode()[0])
dataset['BsmtQual'] = dataset['BsmtQual'].fillna(dataset['BsmtQual'].mode()[0])
dataset['BsmtCond'] = dataset['BsmtCond'].fillna(dataset['BsmtCond'].mode()[0])
dataset['BsmtExposure'] = dataset['BsmtExposure'].fillna(dataset['BsmtExposure'].mode()[0])
dataset['BsmtFinType1'] = dataset['BsmtFinType1'].fillna(dataset['BsmtFinType1'].mode()[0])
dataset['BsmtFinType2'] = dataset['BsmtFinType2'].fillna(dataset['BsmtFinType2'].mode()[0])
dataset['Electrical'] = dataset['Electrical'].fillna(dataset['Electrical'].mode()[0])
dataset['GarageType'] = dataset['GarageType'].fillna(dataset['GarageType'].mode()[0])
dataset['GarageFinish'] = dataset['GarageFinish'].fillna(dataset['GarageFinish'].mode()[0])
dataset['GarageQual'] = dataset['GarageQual'].fillna(dataset['GarageQual'].mode()[0])
dataset['GarageCond'] = dataset['GarageCond'].fillna(dataset['GarageCond'].mode()[0])

testset['MSZoning'] = testset['MSZoning'].fillna(testset['MSZoning'].mode()[0])
testset['Utilities'] = testset['Utilities'].fillna(testset['Utilities'].mode()[0])
testset['Exterior1st'] = testset['Exterior1st'].fillna(testset['Exterior1st'].mode()[0])
testset['Exterior2nd'] = testset['Exterior2nd'].fillna(testset['Exterior2nd'].mode()[0])
testset['MasVnrType'] = testset['MasVnrType'].fillna(testset['MasVnrType'].mode()[0])
testset['BsmtQual'] = testset['BsmtQual'].fillna(testset['BsmtQual'].mode()[0])
testset['BsmtCond'] = testset['BsmtCond'].fillna(testset['BsmtCond'].mode()[0])
testset['BsmtExposure'] = testset['BsmtExposure'].fillna(testset['BsmtExposure'].mode()[0])
testset['BsmtFinType1'] = testset['BsmtFinType1'].fillna(testset['BsmtFinType1'].mode()[0])
testset['BsmtFinType2'] = testset['BsmtFinType2'].fillna(testset['BsmtFinType2'].mode()[0])
testset['KitchenQual'] = testset['KitchenQual'].fillna(testset['KitchenQual'].mode()[0])
testset['Functional'] = testset['Functional'].fillna(testset['Functional'].mode()[0])
testset['GarageType'] = testset['GarageType'].fillna(testset['GarageType'].mode()[0])
testset['GarageFinish'] = testset['GarageFinish'].fillna(testset['GarageFinish'].mode()[0])
testset['GarageQual'] = testset['GarageQual'].fillna(testset['GarageQual'].mode()[0])
testset['GarageCond'] = testset['GarageCond'].fillna(testset['GarageCond'].mode()[0])
testset['SaleType'] = testset['SaleType'].fillna(testset['SaleType'].mode()[0])

Let's save the cleaned and preprocessed data to be used in later sections.

dataset.to_csv('dataset.csv', index=None)
testset.to_csv('testset.csv', index=None)
