==> 06 Artificial Neural Network.ipynb <==
# Artificial Neural Network

Artificial Neural Network with the following configuration reached a root-mean-squared-error of 0.166

import pandas as pd
import numpy as np

import tensorflow as tf

from sklearn.model_selection import train_test_split
from sklearn.metrics import r2_score
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import mean_squared_error

from sklearn.discriminant_analysis import LinearDiscriminantAnalysis

train = pd.read_csv('dataset.csv')
testset = pd.read_csv('testset.csv')
test_all = pd.read_csv('test.csv')

train = train.drop(['Id'], axis=1)
test = testset.drop(['Id'], axis=1)

X = train.iloc[:, :-1].values
y = train.iloc[:, -1].values
test_data = test.iloc[:,:].values

def transform_columns(X, test_data, columns):
    ct = ColumnTransformer(transformers=[('encoder', OneHotEncoder(), columns)], remainder='passthrough')
    # X = np.array(ct.fit_transform(X))
    X = ct.fit_transform(X)
    # test_data = np.array(ct.transform(test_data))
    test_data = ct.transform(test_data)

    return X, test_data

def standardize(X_train, X_test, test_data):
    sc = StandardScaler(with_mean=False)
    X_train = sc.fit_transform(X_train)
    X_test = sc.transform(X_test)
    test_data = sc.transform(test_data)

    return X_train, X_test, test_data
    
def preds_to_file(test_all, preds, filename):
    preds = np.concatenate(preds)
    output = pd.DataFrame({'Id': test_all.Id, 'SalePrice': preds})
    output.to_csv(filename, index=False)
    print(f"File {filename} created!")

def get_ann():
    ann = tf.keras.models.Sequential()
    ann.add(tf.keras.layers.Dense(units=135, activation='relu'))
    ann.add(tf.keras.layers.Dense(units=135, activation='relu'))
    ann.add(tf.keras.layers.Dense(units=300, activation='relu'))
    ann.add(tf.keras.layers.Dense(units=135, activation='relu'))
    ann.add(tf.keras.layers.Dense(units=135, activation='relu'))
    ann.add(tf.keras.layers.Dense(units=1))

    return ann

def perform_discriminant_analysis(n_components, X_train, X_test, test_data):
  lda = LinearDiscriminantAnalysis(n_components=n_components)
  X_train = lda.fit_transform(X_train, y_train)
  X_test = lda.transform(X_test)
  test_data = lda.transform(test_data)

  return X_train, X_test, test_data

 categorical_cols = [1, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 32, 36, 37, 38, 39, 50, 52, 54, 56, 59, 60, 61, 71, 72]
 X, test_data = transform_columns(X, test_data, categorical_cols)
 X = X.toarray()
 test_data = test_data.toarray()

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=1)
X_train, X_test, test_data = standardize(X_train, X_test, test_data)

# X_train, X_test, test_data = perform_discriminant_analysis(200, X_train, X_test, test_data)

print(X_train.shape)
ann = get_ann()
ann.compile(optimizer='adam', loss='mean_squared_error', metrics=[tf.keras.metrics.RootMeanSquaredError()])

ann.fit(X_train, y_train, batch_size=32, epochs=200, validation_data=(X_test, y_test))

y_pred = ann.predict(X_test)
print(r2_score(y_test, y_pred))

rmse = np.sqrt(mean_squared_error(y_test, y_pred))
print("RMSE: %f" % (rmse))

preds = ann.predict(test_data)
preds_to_file(test_all, preds, "ann_reg_01.csv")
