==> 07 Ensemble.ipynb <==
# Ensembling

On this notebook, the best predictions from the models presented before, are combined to one dataframe. Then the mean of predicted prices is calculated for each testset item and used as a new prediction. Ensembling results a root-squared-mean-error of 0.147. So ensembling didn't cause improvement, since the best result was 0.1335.

import os
import pandas as pd

Combine all the model predictions into a single dataframe.

folder = "csvs/"
predictions = pd.DataFrame()
column_count = 0

for index, filename in enumerate(os.listdir(folder)):
  if filename.endswith('.csv'):
    df = pd.read_csv(folder + filename)

    if "Id" not in predictions:
      predictions = df
    else:
      predictions["col_" + str(index)] = df['SalePrice']

    column_count = index + 1


predictions.rename(columns={"SalePrice": "col_0"}, inplace=True)
predictions.head()


# This function gets the "majority vote" of all the model predictions for each row. 
def sale_price(row):
  return row.drop('Id').mean()

# Get the ensembled prediction for each testset item.
predictions['SalePrice'] = predictions.apply(sale_price, axis=1)

predictions.head()

Separate the final ensembled predictions into a dataframe.

ensembled_predictions = predictions[['Id', 'SalePrice']]
ensembled_predictions.head()

ensembled_predictions.to_csv('ensemble_preds_01.csv', index=None)
